No Clutch Inertia
The clutch inertia is moved out of the clutch block and in with the bldc shaft block. Good responses right now with Run 5 signal. Torque reading on SKP looks to be the sum of the load cells. SKP power also looks to be half of the VESC I_motor x Vin (line to line correct needed?)